Add BigtableClientSnippetsFixture
