+++ apis/Google.Cloud.Bigtable.V2/Google.Cloud.Bigtable.V2.IntegrationTests/BigtableFixture.cs
@@ -27,108 +27,23 @@ namespace Google.Cloud.Bigtable.V2.IntegrationTests
     // Note: this does not use CloudProjectFixtureBase as a base class, as an emulator can be used as an alternative.
     // If that pattern becomes more widespread, we can bring that into CloudProjectFixtureBase.
     [CollectionDefinition(nameof(BigtableFixture))]
-    public class BigtableFixture : IDisposable, ICollectionFixture<BigtableFixture>
+    public class BigtableFixture : BigtableFixtureBase, ICollectionFixture<BigtableFixture>
     {
-        private const string EmulatorEnvironmentVariable = "BIGTABLE_EMULATOR_HOST";
-        private const string TestInstanceEnvironmentVariable = "BIGTABLE_TEST_INSTANCE";
-        private const string TestProjectEnvironmentVariable = CloudProjectFixtureBase.TestProjectEnvironmentVariable;
-
         public const string DefaultColumnFamily = "cf1";
         public const string DefaultColumnQualifier = "row_exists";
         public const string DefaultValue = "true";
         public const string OtherColumnFamily = "test_data";
 
-        public ProjectName ProjectName { get; private set; }
-        public InstanceName InstanceName { get; private set; }
-        public TableName TableName { get; private set; }
-
-        public BigtableInstanceAdminClient InstanceAdminClient { get; private set; }
-        public BigtableTableAdminClient TableAdminClient { get; private set; }
-        public BigtableClient TableClient { get; private set; }
-
-        public List<TableName> CreatedTables { get; } = new List<TableName>();
-        public Channel EmulatorChannel { get; }
-
-        public BigtableFixture()
-        {
-            string emulatorHost = Environment.GetEnvironmentVariable(EmulatorEnvironmentVariable);
-
-            string projectId;
-            string instanceId;
-            if (!string.IsNullOrEmpty(emulatorHost))
+        protected override Table CreateDefaultTable() =>
+            new Table
             {
-                projectId = "emulator-test-project";
-                EmulatorChannel = new Channel(emulatorHost, ChannelCredentials.Insecure);
-
-                instanceId = "doesnt-matter";
-            }
-            else
-            {
-                projectId = Environment.GetEnvironmentVariable(TestProjectEnvironmentVariable);
-                if (string.IsNullOrEmpty(projectId))
+                Granularity = Table.Types.TimestampGranularity.Millis,
+                ColumnFamilies =
                 {
-                    throw new InvalidOperationException(
-                        $"Please set either the {EmulatorEnvironmentVariable} or {TestProjectEnvironmentVariable} environment variable before running tests");
+                    { DefaultColumnFamily, new ColumnFamily { GcRule = new GcRule { MaxNumVersions = 3 } } },
+                    { OtherColumnFamily, new ColumnFamily { GcRule = new GcRule { MaxNumVersions = 3 } } }
                 }
-
-                instanceId = Environment.GetEnvironmentVariable(TestInstanceEnvironmentVariable);
-                if (string.IsNullOrEmpty(instanceId))
-                {
-                    throw new InvalidOperationException(
-                        $"Please set the {TestInstanceEnvironmentVariable} environment variable before running non-emulator tests.");
-                }
-            }
-
-            ProjectName = new ProjectName(projectId);
-            InstanceName = new InstanceName(projectId, instanceId);
-
-            Task.Run(InitBigtableInstanceAndTable).Wait();
-        }
-
-        public async Task<TableName> CreateTable()
-        {
-            var tableName = new TableName(ProjectName.ProjectId, InstanceName.InstanceId, Guid.NewGuid().ToString());
-            CreatedTables.Add(tableName);
-            await TableAdminClient.CreateTableAsync(
-                new InstanceName(tableName.ProjectId, tableName.InstanceId),
-                tableName.TableId,
-                new Table
-                {
-                    Granularity = Table.Types.TimestampGranularity.Millis,
-                    ColumnFamilies =
-                    {
-                        { DefaultColumnFamily, new ColumnFamily { GcRule = new GcRule { MaxNumVersions = 3 } } },
-                        { OtherColumnFamily, new ColumnFamily { GcRule = new GcRule { MaxNumVersions = 3 } } }
-                    }
-                });
-            return tableName;
-        }
-
-        private async Task InitBigtableInstanceAndTable()
-        {
-            if (EmulatorChannel == null)
-            {
-                InstanceAdminClient = BigtableInstanceAdminClient.Create();
-                TableAdminClient = BigtableTableAdminClient.Create();
-                TableClient = BigtableClient.Create();
-
-                try
-                {
-                    await InstanceAdminClient.GetInstanceAsync(InstanceName);
-                }
-                catch (RpcException e) when (e.Status.StatusCode == StatusCode.NotFound)
-                {
-                    Assert.True(false, $"The Bigtable instance for testing does not exist: {InstanceName}");
-                }
-            }
-            else
-            {
-                TableAdminClient = BigtableTableAdminClient.Create(EmulatorChannel);
-                TableClient = BigtableClient.Create(EmulatorChannel);
-            }
-
-            TableName = await CreateTable();
-        }
+            };
 
         public async Task<BigtableByteString> InsertRowAsync(
             TableName tableName,
@@ -164,6 +79,7 @@ namespace Google.Cloud.Bigtable.V2.IntegrationTests
             return rowKey;
         }
 
+
         public async Task InsertRowsAsync(
             TableName tableName,
             IEnumerable<BigtableByteString> rowKeys,
@@ -192,21 +108,6 @@ namespace Google.Cloud.Bigtable.V2.IntegrationTests
                                     new BigtableByteString(counter++).Value),
                                 version))).ToArray()).GetResponseEntries();
             Assert.True(entries.All(e => e.Status.Code == (int)Code.Ok));
-        }
-
-        public void Dispose()
-        {
-            if (TableAdminClient != null)
-            {
-                foreach (var tableName in CreatedTables)
-                {
-                    try
-                    {
-                        TableAdminClient.DeleteTable(new DeleteTableRequest { Name = tableName.ToString() });
-                    }
-                    catch { }
-                }
-            }
         }
     }
 }
